fix: emit unescaped bracket keys for CollectionFormat.Indexed
